+++ component/include/pl_server.h
@@ -17,7 +17,6 @@ public:
 
   /// @brief Creates a server
   Server();
-  virtual ~Server() = default;
 
   /// @brief Enables the server
   /// @return error code